Refactoring the code in separate files and changing some of the file names.
