+++ lib/pembase64.c
@@ -28,6 +28,7 @@
 #include <stdio.h>
 #include "picotls.h"
 #include "picotls/minicrypto.h"
+#include "picotls/asn1.h"
 
 static char ptls_base64_alphabet[] = {
     'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M',
@@ -254,397 +255,6 @@ int ptls_base64_decode(const char * text, ptls_base64_decode_state_t * state, pt
 }
 
 /*
- * Basic ASN1 validation and optional print-out
- */
-
-static char const * asn1_type_classes[4] = {
-	"Universal",
-	"Application",
-	"Context-specific",
-	"Private"
-};
-
-static char const * asn1_universal_types[] = {
-	"End-of-Content",
-	"BOOLEAN",
-	"INTEGER",
-	"BIT STRING",
-	"OCTET STRING",
-	"NULL",
-	"OBJECT IDENTIFIER",
-	"Object Descriptor",
-	"EXTERNAL",
-	"REAL",
-	"ENUMERATED",
-	"EMBEDDED PDV",
-	"UTF8String",
-	"RELATIVE-OID",
-	"Reserved (16)",
-	"Reserved (17)",
-	"SEQUENCE",
-	"SET",
-	"NumericString",
-	"PrintableString",
-	"T61String",
-	"VideotexString",
-	"IA5String",
-	"UTCTime",
-	"GeneralizedTime",
-	"GraphicString",
-	"VisibleString",
-	"GeneralString",
-	"UniversalString",
-	"CHARACTER STRING",
-	"BMPString"
-};
-
-
-/* For debugging
-*/
-
-static void data_dump(const uint8_t * bytes, size_t length, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	size_t byte_index = 0;
-
-	while (byte_index < length)
-	{
-		log_ctx->fn(log_ctx->ctx, "%06x ", (uint32_t) byte_index);
-		for (size_t i = 0; i < 32 && byte_index < length; i++, byte_index++)
-		{
-			log_ctx->fn(log_ctx->ctx, "%02x", bytes[byte_index]);
-			if ((i & 3) == 3)
-			{
-				log_ctx->fn(log_ctx->ctx, " ");
-			}
-		}
-		log_ctx->fn(log_ctx->ctx, "\n");
-	}
-}
-
-static size_t nb_asn1_universal_types = sizeof(asn1_universal_types) / sizeof(char const *);
-
-static void ptls_asn1_print_indent(int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	for (int indent = 0; indent <= level; indent++)
-	{
-		log_ctx->fn(log_ctx->ctx, "   ");
-	}
-}
-
-
-static size_t ptls_asn1_error_message(char const * error_label, size_t bytes_max, size_t byte_index, 
-	int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	if (log_ctx != NULL)
-	{
-		ptls_asn1_print_indent(level, log_ctx);
-		log_ctx->fn(log_ctx->ctx, "Error: %s (near position: %d (0x%x) out of %d)", 
-			error_label, (int) byte_index, (uint32_t) byte_index, (int)bytes_max);
-	}
-	return byte_index;
-}
-
-static void ptls_asn1_dump_content(const uint8_t * bytes, size_t bytes_max, size_t byte_index, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	if (log_ctx != NULL && bytes_max > byte_index)
-	{
-		size_t nb_bytes = bytes_max - byte_index;
-
-		log_ctx->fn(log_ctx->ctx, " ");
-
-		for (size_t i = 0; i < 16 && i < nb_bytes; i++)
-		{
-			log_ctx->fn(log_ctx->ctx, "%02x", bytes[byte_index + i]);
-		}
-
-		if (nb_bytes > 16)
-		{
-			log_ctx->fn(log_ctx->ctx, "...");
-		}
-	}
-}
-
-size_t ptls_asn1_read_type(const uint8_t * bytes, size_t bytes_max,
-	int * structure_bit, int * type_class, uint32_t * type_number,
-	int * decode_error, int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	/* Get the type byte */
-	size_t byte_index = 1;
-	uint8_t first_byte = bytes[0];
-	*structure_bit = (first_byte >> 5) & 1;
-	*type_class = (first_byte >> 6) & 3;
-	*type_number = first_byte & 31;
-
-	if (*type_number == 31)
-	{
-		uint32_t long_type = 0;
-		const uint32_t type_number_limit = 0x07FFFFFFF;
-		int next_byte;
-		int end_found = 0;
-
-		while (byte_index < bytes_max && long_type <= type_number_limit) {
-			next_byte = bytes[byte_index++];
-			long_type <<= 7;
-			long_type |= next_byte & 127;
-			if ((next_byte & 128) == 0)
-			{
-				end_found = 1;
-				break;
-			}
-		}
-
-		if (end_found)
-		{
-			*type_number = long_type;
-		}
-		else
-		{
-			/* This is an error */
-			byte_index = ptls_asn1_error_message("Incorrect type coding", bytes_max, byte_index,
-				level, log_ctx);
-			*decode_error = PTLS_ERROR_BER_MALFORMED_TYPE;
-		}
-	}
-
-	return byte_index;
-}
-
-void ptls_asn1_print_type(int type_class, uint32_t type_number, int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	/* Print the type */
-	ptls_asn1_print_indent(level, log_ctx);
-	if (type_class == 0 && type_number < nb_asn1_universal_types)
-	{
-		log_ctx->fn(log_ctx->ctx, "%s", asn1_universal_types[type_number]);
-	}
-	else if (type_class == 2)
-	{
-		log_ctx->fn(log_ctx->ctx, "[%d]", type_number);
-	}
-	else
-	{
-		log_ctx->fn(log_ctx->ctx, "%s[%d]", asn1_type_classes[type_class], type_number);
-	}
-}
-
-size_t ptls_asn1_read_length(const uint8_t * bytes, size_t bytes_max, size_t byte_index,
-	uint32_t * length, int * indefinite_length, size_t * last_byte,
-	int * decode_error, int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	int length_of_length = 0;
-
-	*indefinite_length = 0;
-	*length = 0;
-	*last_byte = bytes_max;
-
-	if (byte_index < bytes_max)
-	{
-		*length = bytes[byte_index++];
-		if ((*length & 128) != 0)
-		{
-			length_of_length = *length & 127;
-			*length = 0;
-
-			if (byte_index + length_of_length >= bytes_max)
-			{
-				/* This is an error */
-				byte_index = ptls_asn1_error_message("Incorrect length coding", bytes_max, byte_index,
-					level, log_ctx);
-				*decode_error = PTLS_ERROR_BER_MALFORMED_LENGTH;
-			}
-			else
-			{
-				for (int i = 0; i < length_of_length && byte_index < bytes_max; i++)
-				{
-					*length <<= 8;
-					*length |= bytes[byte_index++];
-				}
-
-				if (length_of_length == 0)
-				{
-					*last_byte = bytes_max;
-					*indefinite_length = 1;
-				}
-				else
-				{
-					*last_byte = byte_index + *length;
-				}
-			}
-		}
-		else
-		{
-			*last_byte = byte_index + *length;
-		}
-
-		if (*decode_error == 0)
-		{
-			/* TODO: verify that the length makes sense */
-			if (*last_byte > bytes_max)
-			{
-				byte_index = ptls_asn1_error_message("Length larger than message", bytes_max, byte_index,
-					level, log_ctx);
-				*decode_error = PTLS_ERROR_BER_EXCESSIVE_LENGTH;
-			}
-		}
-	}
-
-	return byte_index;
-}
-
-size_t ptls_asn1_get_expected_type_and_length(const uint8_t * bytes, size_t bytes_max, size_t byte_index,
-	uint8_t expected_type, uint32_t * length, int * indefinite_length, size_t * last_byte,
-	int * decode_error, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	int is_indefinite = 0;
-
-	/* Check that the expected type is present */
-	if (bytes[byte_index] != expected_type)
-	{
-		byte_index = ptls_asn1_error_message("Unexpected type", bytes_max, byte_index,
-			0, log_ctx);
-		*decode_error = PTLS_ERROR_INCORRECT_PEM_SYNTAX;
-	}
-	else
-	{
-		/* get length of element */
-		byte_index++;
-		byte_index = ptls_asn1_read_length(bytes, bytes_max, byte_index,
-			length, &is_indefinite, last_byte, decode_error, 0, log_ctx);
-
-		if (indefinite_length != NULL)
-		{
-			*indefinite_length = is_indefinite;
-		}
-		else if (is_indefinite)
-		{
-			byte_index = ptls_asn1_error_message("Incorrect length for DER", bytes_max, byte_index,
-				0, log_ctx);
-			*decode_error = PTLS_ERROR_DER_INDEFINITE_LENGTH;
-		}
-	}
-
-	return byte_index;
-}
-
-size_t ptls_asn1_validation_recursive(const uint8_t * bytes, size_t bytes_max, 
-    int * decode_error, int level, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-    /* Get the type byte */
-	int structure_bit = 0;
-	int type_class = 0;
-    uint32_t type_number = 0;
-    uint32_t length = 0;
-    int indefinite_length = 0;
-    size_t last_byte = 0;
-	/* Decode the type */
-	size_t byte_index = ptls_asn1_read_type(bytes, bytes_max, &structure_bit, &type_class, &type_number,
-		decode_error, level, log_ctx);
-
-	if (*decode_error == 0 && log_ctx != NULL)
-	{
-		ptls_asn1_print_type(type_class, type_number, level, log_ctx);
-	}
-
-
-    /* Get the length */
-	byte_index = ptls_asn1_read_length(bytes, bytes_max, byte_index,
-		&length, &indefinite_length, &last_byte,
-		decode_error, level, log_ctx);
-
-	if (last_byte <= bytes_max)
-	{
-		if (structure_bit)
-		{
-			/* If structured, recurse on a loop */
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, " {\n");
-			}
-
-			while (byte_index < last_byte)
-			{
-				if (indefinite_length != 0 &&
-					bytes[byte_index] == 0)
-				{
-					if (byte_index + 2 > bytes_max ||
-						bytes[byte_index + 1] != 0)
-					{
-						byte_index = ptls_asn1_error_message("EOC: unexpected end of content",
-							bytes_max, byte_index, level + 1, log_ctx);
-
-						*decode_error = PTLS_ERROR_BER_UNEXPECTED_EOC;
-					}
-					else
-					{
-						if (log_ctx != NULL)
-						{
-							ptls_asn1_print_indent(level, log_ctx);
-							log_ctx->fn(log_ctx->ctx, "EOC\n");
-						}
-						byte_index += 2;
-						break;
-					}
-				}
-				else
-				{
-					byte_index += ptls_asn1_validation_recursive(
-						bytes + byte_index, last_byte - byte_index,
-						decode_error, level + 1, log_ctx);
-
-					if (*decode_error)
-					{
-						byte_index = bytes_max;
-						break;
-					}
-				}
-
-				if (log_ctx != NULL)
-				{
-					if (byte_index < last_byte)
-					{
-						log_ctx->fn(log_ctx->ctx, ",");
-					}
-					log_ctx->fn(log_ctx->ctx, "\n");
-				}
-			}
-
-
-			if (log_ctx != NULL)
-			{
-				ptls_asn1_print_indent(level, log_ctx);
-				log_ctx->fn(log_ctx->ctx, "}");
-			}
-		}
-		else
-		{
-			ptls_asn1_dump_content(bytes, last_byte, byte_index, log_ctx);
-			byte_index = last_byte;
-		}
-	}
-
-    return byte_index;
-}
-
-int ptls_asn1_validation(const uint8_t * bytes, size_t length, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	int decode_error = 0;
-	size_t decoded = ptls_asn1_validation_recursive(bytes, length,
-		&decode_error, 0, log_ctx);
-
-	if (decode_error == 0 && decoded < length)
-	{
-		decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "Type too short, %d bytes only out of %d\n",
-				(int)decoded, (int)length);
-		}
-	}
-
-	return decode_error;
-}
-
-/*
  * Reading a PEM file, to get an object:
  *
  * - Find first object, get the object name.
@@ -699,7 +309,7 @@ static int ptls_compare_separator_line(const char * line, const char* begin_or_e
     return ret;
 }
 
-static int ptls_get_pem_object(FILE * F, const char * label, ptls_buffer_t *buf, ptls_minicrypto_log_ctx_t * log_ctx)
+static int ptls_get_pem_object(FILE * F, const char * label, ptls_buffer_t *buf)
 {
     int ret = PTLS_ERROR_PEM_LABEL_NOT_FOUND;
     char line[256];
@@ -737,24 +347,11 @@ static int ptls_get_pem_object(FILE * F, const char * label, ptls_buffer_t *buf,
         }
     }
 
-	if (ret == 0)
-	{
-		ret = ptls_asn1_validation(buf->base, buf->off, log_ctx);
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "\n");
-		}
-
-		if (ret != 0)
-		{
-			data_dump(buf->base, buf->off, log_ctx);
-		}
-	}
     return ret;
 }
 
-int ptls_pem_get_objects(char const * pem_fname, const char * label, 
-	ptls_iovec_t ** list, size_t list_max, size_t * nb_objects, ptls_minicrypto_log_ctx_t * log_ctx)
+int ptls_minicrypto_get_pem_objects(char const * pem_fname, const char * label, 
+	ptls_iovec_t ** list, size_t list_max, size_t * nb_objects)
 {
     FILE * F;
     int ret = 0;
@@ -784,7 +381,7 @@ int ptls_pem_get_objects(char const * pem_fname, const char * label,
 
 			ptls_buffer_init(&buf, "", 0);
 
-            ret = ptls_get_pem_object(F, label, &buf, log_ctx);
+            ret = ptls_get_pem_object(F, label, &buf);
 
             if (ret == 0)
             {
@@ -821,641 +418,4 @@ int ptls_pem_get_objects(char const * pem_fname, const char * label,
 
     return ret;
 }
-
-int ptls_pem_get_certificates(char const * pem_fname, ptls_iovec_t ** list, size_t list_max, 
-	size_t * nb_certs, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-    return ptls_pem_get_objects(pem_fname, "CERTIFICATE", list, list_max, nb_certs, log_ctx);
-}
-
-
-size_t ptls_pem_decode_private_key(
-	ptls_asn1_pkcs8_private_key_t * pkey,
-	int * decode_error, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	uint8_t * bytes = pkey->vec.base;
-	size_t bytes_max = pkey->vec.len;
-
-	/* read the ASN1 messages */
-	size_t byte_index = 0;
-	uint32_t seq0_length = 0;
-	size_t last_byte0;
-	uint32_t seq1_length = 0;
-	size_t last_byte1 = 0;
-	uint32_t oid_length;
-	size_t last_oid_byte;
-	uint32_t key_data_length;
-	size_t key_data_last;
-
-
-	/* start with sequence */
-	byte_index = ptls_asn1_get_expected_type_and_length(
-		bytes, bytes_max, byte_index, 0x30,
-		&seq0_length, NULL, &last_byte0, decode_error, log_ctx);
-
-	if (decode_error == 0 && bytes_max != last_byte0)
-	{
-		byte_index = ptls_asn1_error_message("Length larger than message", bytes_max, byte_index,
-			0, log_ctx);
-		*decode_error = PTLS_ERROR_BER_EXCESSIVE_LENGTH;
-	}
-
-	if (*decode_error == 0)
-	{
-		/* get first component: version, INTEGER, expect value 0 */
-		if (byte_index + 3 > bytes_max)
-		{
-			byte_index = ptls_asn1_error_message("Cannot find key version", bytes_max, byte_index,
-				0, log_ctx);
-			*decode_error = PTLS_ERROR_INCORRECT_PEM_KEY_VERSION;
-		}
-		else if (bytes[byte_index] != 0x02 ||
-			bytes[byte_index + 1] != 0x01 ||
-			bytes[byte_index + 2] != 0x00)
-		{
-			*decode_error = PTLS_ERROR_INCORRECT_PEM_KEY_VERSION;
-			byte_index = ptls_asn1_error_message("Incorrect PEM Version", bytes_max, byte_index,
-				0, log_ctx);
-		}
-		else
-		{
-			byte_index += 3;
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "   Version = 1,\n");
-			}
-		}
-	}
-
-	if (*decode_error == 0)
-	{
-		/* open embedded sequence */
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, bytes_max, byte_index, 0x30,
-			&seq1_length, NULL, &last_byte1, decode_error, log_ctx);
-	}
-
-	if (*decode_error == 0)
-	{
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "   Algorithm Identifier:\n");
-		}
-		/* get length of OID */
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, last_byte1, byte_index, 0x06,
-			&oid_length, NULL, &last_oid_byte, decode_error, log_ctx);
-
-		if (*decode_error == 0)
-		{
-			if (log_ctx != NULL)
-			{
-				/* print the OID value */
-				log_ctx->fn(log_ctx->ctx, "      Algorithm:");
-				ptls_asn1_dump_content(bytes + byte_index, oid_length, 0, log_ctx);
-				log_ctx->fn(log_ctx->ctx, ",\n");
-			}
-			pkey->algorithm_index = byte_index;
-			pkey->algorithm_length = oid_length;
-			byte_index += oid_length;
-		}
-	}
-
-	if (*decode_error == 0)
-	{
-		/* get parameters, ANY */
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "      Parameters:\n");
-		}
-
-		pkey->parameters_index = byte_index;
-
-		pkey->parameters_length = ptls_asn1_validation_recursive(bytes + byte_index,
-			last_byte1 - byte_index, decode_error, 2, log_ctx);
-
-		byte_index += pkey->parameters_length;
-
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "\n");
-		}
-		/* close sequence */
-		if (byte_index != last_byte1)
-		{
-			byte_index = ptls_asn1_error_message("Length larger than element", bytes_max, byte_index,
-				2, log_ctx);
-			*decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-		}
-	}
-
-	/* get octet string, key */
-	if (*decode_error == 0)
-	{
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, last_byte0, byte_index, 0x04,
-			&key_data_length, NULL, &key_data_last, decode_error, log_ctx);
-
-		if (*decode_error == 0)
-		{
-			pkey->key_data_index = byte_index;
-			pkey->key_data_length = key_data_length;
-			byte_index += key_data_length;
-
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "   Key data (%d bytes):\n", key_data_length);
-
-				(void)ptls_asn1_validation_recursive(bytes + pkey->key_data_index,
-					key_data_length, decode_error, 1, log_ctx);
-				log_ctx->fn(log_ctx->ctx, "\n");
-			}
-		}
-	}
-
-	if (*decode_error == 0 && byte_index != last_byte0)
-	{
-		byte_index = ptls_asn1_error_message("Length larger than element", bytes_max, byte_index,
-			0, log_ctx);
-		*decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-	}
-
-	if (log_ctx != NULL)
-	{
-		log_ctx->fn(log_ctx->ctx, "\n");
-	}
-
-	return byte_index;
-}
-
-int ptls_pem_parse_private_key(char const * pem_fname,
-	ptls_asn1_pkcs8_private_key_t * pkey, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	size_t nb_keys = 0;
-	ptls_iovec_t * list = &pkey->vec;
-	int ret = ptls_pem_get_objects(pem_fname, "PRIVATE KEY", &list, 1, &nb_keys, NULL);
-
-	if (ret == 0)
-	{
-		if (nb_keys != 1)
-		{
-			ret = PTLS_ERROR_PEM_LABEL_NOT_FOUND;
-		}
-	}
-
-	if (ret == 0 && nb_keys == 1)
-	{
-		int decode_error = 0;
-		size_t byte_index = 0;
-
-
-
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "\nFound PRIVATE KEY, length = %d bytes\n", (int)pkey->vec.len);
-		}
-
-		byte_index = ptls_pem_decode_private_key(pkey, &decode_error, log_ctx);
-
-		if (decode_error != 0)
-		{
-			ret = decode_error;
-		}
-	}
-
-	return ret;
-}
-
-const uint8_t ptls_asn1_algorithm_ecdsa[] = { 
-	0x2a, 0x86, 0x48, 0xce, 0x3d, 0x02, 0x01 };
-
-const uint8_t ptls_asn1_curve_secp256r1[] = {
-	0x2a, 0x86, 0x48, 0xce, 0x3d, 0x03, 0x01, 0x07};
-
-
-int ptls_set_ecdsa_private_key(ptls_context_t * ctx,
-	ptls_asn1_pkcs8_private_key_t * pkey, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	uint8_t * bytes = pkey->vec.base + pkey->parameters_index;
-	size_t bytes_max = pkey->parameters_length;
-	size_t byte_index = 0;
-	uint8_t * curve_id = NULL;
-	uint32_t curve_id_length = 0;
-	int decode_error = 0;
-	uint32_t seq_length;
-	size_t last_byte = 0;
-	uint8_t * ecdsa_key_data = NULL;
-	uint32_t ecdsa_key_data_length = 0;
-	size_t  ecdsa_key_data_last = 0;
-
-	/* We expect the parameters to include just the curve ID */
-
-	byte_index = ptls_asn1_get_expected_type_and_length(
-		bytes, bytes_max, byte_index, 0x06,
-		&curve_id_length, NULL, &last_byte, &decode_error, log_ctx);
-
-	if (decode_error == 0 && bytes_max != last_byte)
-	{
-		byte_index = ptls_asn1_error_message("Length larger than parameters", bytes_max, byte_index,
-			0, log_ctx);
-		decode_error = PTLS_ERROR_BER_EXCESSIVE_LENGTH;
-	}
-
-	if (decode_error == 0)
-	{
-		curve_id = bytes + byte_index;
-
-		if (log_ctx != NULL)
-		{
-			/* print the OID value */
-			log_ctx->fn(log_ctx->ctx, "Curve: ");
-			ptls_asn1_dump_content(curve_id, curve_id_length, 0, log_ctx);
-			log_ctx->fn(log_ctx->ctx, "\n");
-		}
-	}
-
-	/* We expect the key data to follow the ECDSA structure per RFC 5915 */
-	bytes = pkey->vec.base + pkey->key_data_index;
-	bytes_max = pkey->key_data_length;
-	byte_index = 0;
-
-	/* decode the wrapping sequence */
-	if (decode_error == 0)
-	{
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, bytes_max, byte_index, 0x30,
-			&seq_length, NULL, &last_byte, &decode_error, log_ctx);
-	}
-
-	if (decode_error == 0 && bytes_max != last_byte)
-	{
-		byte_index = ptls_asn1_error_message("Length larger than key data", bytes_max, byte_index,
-			0, log_ctx);
-		decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-	}
-
-	/* verify and skip the version number 1 */
-	if (decode_error == 0)
-	{
-		/* get first component: version, INTEGER, expect value 0 */
-		if (byte_index + 3 > bytes_max)
-		{
-			byte_index = ptls_asn1_error_message("Cannot find ECDSA Key Data Version", bytes_max, byte_index,
-				0, log_ctx);
-			decode_error = PTLS_ERROR_INCORRECT_PEM_SYNTAX;
-		}
-		else if (bytes[byte_index] != 0x02 ||
-			bytes[byte_index + 1] != 0x01 ||
-			bytes[byte_index + 2] != 0x01)
-		{
-			decode_error = PTLS_ERROR_INCORRECT_PEM_ECDSA_KEY_VERSION;
-			byte_index = ptls_asn1_error_message("Incorrect ECDSA Key Data Version", bytes_max, byte_index,
-				0, log_ctx);
-		}
-		else
-		{
-			byte_index += 3;
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "ECDSA Version = 1,\n");
-			}
-		}
-	}
-
-	/* obtain the octet string that contains the ECDSA private key */
-	if (decode_error == 0)
-	{
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, last_byte, byte_index, 0x04,
-			&ecdsa_key_data_length, NULL, &ecdsa_key_data_last, &decode_error, log_ctx);
-
-		if (decode_error == 0)
-		{
-			ecdsa_key_data = bytes + byte_index;
-		}
-	}
-
-	/* If everything is fine, associate the ECDSA key with the context */
-	if (curve_id_length == sizeof(ptls_asn1_curve_secp256r1) && curve_id != NULL &&
-		memcmp(curve_id, ptls_asn1_curve_secp256r1, sizeof(ptls_asn1_curve_secp256r1)) == 0)
-	{
-		if (SECP256R1_PRIVATE_KEY_SIZE != ecdsa_key_data_length)
-		{
-			decode_error = PTLS_ERROR_INCORRECT_PEM_ECDSA_KEYSIZE;
-			if (log_ctx != NULL)
-			{
-				/* print the OID value */
-				log_ctx->fn(log_ctx->ctx, "Wrong SECP256R1 key length, %d instead of %d.\n",
-					ecdsa_key_data_length, SECP256R1_PRIVATE_KEY_SIZE);
-			}
-		}
-		else
-		{
-			ptls_minicrypto_secp256r1sha256_sign_certificate_t * minicrypto_sign_certificate;
-
-			minicrypto_sign_certificate =
-				(ptls_minicrypto_secp256r1sha256_sign_certificate_t *)malloc(
-					sizeof(ptls_minicrypto_secp256r1sha256_sign_certificate_t));
-
-			if (minicrypto_sign_certificate == NULL)
-			{
-				decode_error = PTLS_ERROR_NO_MEMORY;
-			}
-			else
-			{
-				memset(minicrypto_sign_certificate, 0, 
-					sizeof(ptls_minicrypto_secp256r1sha256_sign_certificate_t));
-				decode_error = ptls_minicrypto_init_secp256r1sha256_sign_certificate(
-					minicrypto_sign_certificate, ptls_iovec_init(ecdsa_key_data, ecdsa_key_data_length));
-			}
-			if (decode_error == 0)
-			{
-				ctx->sign_certificate = &minicrypto_sign_certificate->super;
-
-				if (log_ctx != NULL)
-				{
-					/* print the OID value */
-					log_ctx->fn(log_ctx->ctx, "Initialized SECP512R1 signing key with %d bytes.\n",
-						ecdsa_key_data_length);
-				}
-			}
-			else if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "SECP512R1 init with %d bytes returns %d.\n", ecdsa_key_data_length, decode_error);
-			}
-		}
-	}
-	else
-	{
-		decode_error = PTLS_ERROR_INCORRECT_PEM_ECDSA_CURVE;
-		if (log_ctx != NULL)
-		{
-			/* print the OID value */
-			log_ctx->fn(log_ctx->ctx, "Curve is not supported for signatures.\n");
-		}
-	}
-
-	return decode_error;
-}
-
-int ptls_set_private_key(ptls_context_t * ctx, char const * pem_fname, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	ptls_asn1_pkcs8_private_key_t pkey = { {0} };
-	int ret = ptls_pem_parse_private_key(pem_fname, &pkey, log_ctx);
-
-	/* Check that this is the expected key type.
-	 * At this point, the minicrypto library only supports ECDSA keys.
-	 * In theory, we could add support for RSA keys at some point.
-	 */
-	if (ret == 0)
-	{
-		if (pkey.algorithm_length == sizeof(ptls_asn1_algorithm_ecdsa) &&
-			memcmp(pkey.vec.base + pkey.algorithm_index,
-				ptls_asn1_algorithm_ecdsa, sizeof(ptls_asn1_algorithm_ecdsa)) == 0)
-		{
-			ret = ptls_set_ecdsa_private_key(ctx, &pkey, log_ctx);
-		}
-		else
-		{
-			ret = -1;
-		}
-	}
-
-	return ret;
-}
-
-#if 0
-int ptls_pem_get_private_key(char const * pem_fname, ptls_iovec_t * vec, 
-	ptls_minicrypto_log_ctx_t * log_ctx)
-{
-	size_t nb_keys = 0;
-	int ret = ptls_pem_get_objects(pem_fname, "PRIVATE KEY", &vec, 1, &nb_keys, NULL);
-
-	if (ret == 0)
-	{
-		if (nb_keys != 1)
-		{
-			ret = PTLS_ERROR_PEM_LABEL_NOT_FOUND;
-		}
-	}
-	if (ret == 0 && nb_keys == 1)
-	{
-		/* read the ASN1 messages */
-		size_t byte_index = 0;
-		uint8_t * bytes = vec->base;
-		size_t bytes_max = vec->len;
-		int decode_error = 0;
-		uint32_t seq0_length = 0;
-		size_t last_byte0;
-		uint32_t seq1_length = 0;
-		size_t last_byte1 = 0;
-		uint32_t oid_length;
-		size_t last_oid_byte;
-		uint32_t key_data_length;
-		size_t key_data_last;
-
-
-		if (log_ctx != NULL)
-		{
-			log_ctx->fn(log_ctx->ctx, "\nFound PRIVATE KEY, length = %d bytes\n", (int)bytes_max);
-		}
-
-		/* start with sequence */
-		byte_index = ptls_asn1_get_expected_type_and_length(
-			bytes, bytes_max, byte_index, 0x30,
-			&seq0_length, NULL, &last_byte0, &decode_error, log_ctx);
-
-		if (decode_error == 0 && bytes_max != last_byte0)
-		{
-			byte_index = ptls_asn1_error_message("Length larger than message", bytes_max, byte_index,
-					0, log_ctx);
-			decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-		}
-
-		if (decode_error == 0)
-		{
-			/* get first component: version, INTEGER, expect value 0 */
-			if (byte_index + 3 > bytes_max)
-			{
-				byte_index = ptls_asn1_error_message("Incorrect length for DER", bytes_max, byte_index,
-					0, log_ctx);
-				decode_error = PTLS_ERROR_INCORRECT_PEM_KEY_VERSION;
-			}
-			else if (bytes[byte_index] != 0x02 ||
-				bytes[byte_index + 1] != 0x01 ||
-				bytes[byte_index + 2] != 0x00)
-			{
-				decode_error = PTLS_ERROR_INCORRECT_PEM_KEY_VERSION;
-				byte_index = ptls_asn1_error_message("Incorrect PEM Version", bytes_max, byte_index,
-					0, log_ctx);
-			}
-			else
-			{
-				byte_index += 3; 
-				if (log_ctx != NULL)
-				{
-					log_ctx->fn(log_ctx->ctx, "   Version = 1,\n");
-				}
-			}
-		}
-
-		if (decode_error == 0)
-		{
-			/* open embedded sequence */
-			byte_index = ptls_asn1_get_expected_type_and_length(
-				bytes, bytes_max, byte_index, 0x30,
-				&seq1_length, NULL, &last_byte1, &decode_error, log_ctx);
-		}
-
-		if (decode_error == 0)
-		{
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "   Algorithm Identifier:\n");
-			}
-			/* get length of OID */
-			byte_index = ptls_asn1_get_expected_type_and_length(
-				bytes, last_byte1, byte_index, 0x06,
-				&oid_length, NULL, &last_oid_byte, &decode_error, log_ctx);
-			
-			if (decode_error == 0)
-			{
-				if (log_ctx != NULL)
-				{
-					/* print the OID value */
-					log_ctx->fn(log_ctx->ctx, "      Algorithm:");
-					ptls_asn1_dump_content(bytes + byte_index, oid_length, 0, log_ctx);
-					log_ctx->fn(log_ctx->ctx, ",\n");
-				}
-				byte_index += oid_length;
-			}
-		}
-
-		if (decode_error == 0)
-		{
-			/* get parameters, ANY */
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "      Parameters:\n");
-			}
-			byte_index += ptls_asn1_validation_recursive(bytes + byte_index, 
-				last_byte1 - byte_index, &decode_error, 2, log_ctx);
-			if (log_ctx != NULL)
-			{
-				log_ctx->fn(log_ctx->ctx, "\n");
-			}
-			/* close sequence */
-			if (byte_index != last_byte1)
-			{
-				byte_index = ptls_asn1_error_message("Length larger than element", bytes_max, byte_index,
-					2, log_ctx);
-				decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-			}
-		}
-
-		/* get octet string, key */
-		if (decode_error == 0)
-		{
-			byte_index = ptls_asn1_get_expected_type_and_length(
-				bytes, bytes_max, byte_index, 0x04,
-				&key_data_length, NULL, &key_data_last, &decode_error, log_ctx);
-
-			if (decode_error == 0)
-			{
-				if (log_ctx != NULL)
-				{
-					log_ctx->fn(log_ctx->ctx, "   Key data (%d bytes):\n", key_data_length);
-				}
-				/* print octet string as ASN.1 component */
-				if (byte_index != last_byte1)
-				{
-					byte_index += ptls_asn1_validation_recursive(bytes + byte_index,
-						key_data_length, &decode_error, 1, log_ctx);
-					if (log_ctx != NULL)
-					{
-						log_ctx->fn(log_ctx->ctx, "\n");
-					}
-				}
-			}
-		}
-		if (decode_error == 0 && byte_index != last_byte0)
-		{
-			byte_index = ptls_asn1_error_message("Length larger than element", bytes_max, byte_index,
-				0, log_ctx);
-			decode_error = PTLS_ERROR_BER_ELEMENT_TOO_SHORT;
-		}
-
-		if (decode_error != 0)
-		{
-			ret = decode_error;
-		}
-	}
-	return ret;
-}
-#endif
-/*
-IMO it could be implemented in either of the three ways :
-
-a) a function that accepts `ptls_context_t *` and a filename as
-arguments.The function will populate `ptls_context_t::certificates`.
-A function that disposes of the memory allocated for the certificates
-stored in the context needs to be defined as well.
-b) a function that accepts `ptls_context_t *` and `FILE *` as
-arguments.The function will populate `ptls_context_t::certificates`.
-A function that disposes of the memory allocated for the certificates
-stored in the context needs to be defined as well.
-c) a function that accepts `FILE *` and returns at most one
-certificate(in type ptls_iovec_t, with `ptls_iovec_t::base` being
-    allocated using malloc).User should call the function repeatedly
-    until it returns an EOF.It is the users' responsibility to setup
-    `ptls_context_t::certificates` by using the values returned the
-    function.
-
-    I do not have a strong preference between the three, though it might
-    make sense to implement(c) and optionally provide a wrapper in the
-    style of either(a) or (b).
-*/
-#define PTLS_MAX_CERTS_IN_CONTEXT 16
-
-int ptls_set_context_certificates(ptls_context_t * ctx, 
-    char * cert_pem_file, ptls_minicrypto_log_ctx_t * log_ctx)
-{
-    int ret = 0;
-
-    ctx->certificates.list = (ptls_iovec_t *)
-        malloc(PTLS_MAX_CERTS_IN_CONTEXT * sizeof(ptls_iovec_t));
-
-    if (ctx->certificates.list == NULL)
-    {
-        ret = PTLS_ERROR_NO_MEMORY;
-    }
-    else
-    {
-        ret = ptls_pem_get_objects(cert_pem_file, "CERTIFICATE",
-            &ctx->certificates.list, PTLS_MAX_CERTS_IN_CONTEXT, &ctx->certificates.count, log_ctx);
-    }
-
-    return ret;
-}
-
-
-/*
-> 2) What is the proper API to push signing keys in mini crypto.
-
-IMO it should either be :
-
-a) a function that initializes
-`ptls_minicrypto_secp256r1sha256_sign_certificate_t` (much like
-    ptls_minicrypto_init_secp256r1sha256_sign_certificate), taking address
-    of the object and a filename of the private key as arguments
-    b) a function that initializes
-    `ptls_minicrypto_secp256r1sha256_sign_certificate_t` (much like
-        ptls_minicrypto_init_secp256r1sha256_sign_certificate), taking address
-    of the object and `FILE *` of the private key as arguments
-    c) a function that reads a private key from file(either specified by
-    a filename or a file pointer) and returns it as `ptls_iovec_t`
-
-    For the matter, I think that having(c), as well as optionally having
-    either(a) or (b)might make sense.
-*/
