+++ picotlsvs/picotls/picotls.vcxproj.filters
@@ -84,6 +84,12 @@
     <ClInclude Include="..\..\deps\cifra\src\tassert.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\include\picotls\minicrypto.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\include\picotls\asn1.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\lib\openssl.c">
@@ -102,6 +108,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\..\lib\pembase64.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\lib\asn1.c">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
